Server/requirements.txt
fastapi
pydantic[email]
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
